Working playback again after refactoring
