+++ app/build.gradle
@@ -13,7 +13,7 @@ android {
         minSdk 29
         targetSdk 33
         versionCode 1
-        versionName "1.1.0"
+        versionName "1.2.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -45,7 +45,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     // 5GMAG
-    implementation 'com.fivegmag:a5gmscommonlibrary:1.1.0'
+    implementation 'com.fivegmag:a5gmscommonlibrary:1.2.0'
 
     // Retrofit
     def retrofit_version = "2.9.0"
